refactor: SleepTrainWidgetAttributes 분리하여 public으로
